Split stat.cpp into smaller pieces
